requirements.txt
openai
langchain
langchain_community
langchain-openai
annoy
aiohttp
bs4
python-dotenv
certifi
pypdf
tiktoken
pydantic
streamlit
arize-phoenix
arize
arize-otel 
openai 
openinference-instrumentation-langchain>=0.1.15
opentelemetry-sdk 
opentelemetry-exporter-otlp
chromadb